+++ metabot/functions/src/commands/index.ts
@@ -1,0 +1,11 @@
+import * as yargs from 'yargs'
+import * as execCommand from './exec'
+import * as inviteCommand from './invite'
+
+function instHandler(command): yargs.CommandModule {
+  command.handler = (argv) => { /* noop */ }
+  return command
+}
+
+export const exec = instHandler(execCommand)
+export const invite = instHandler(inviteCommand)
