+++ metabot/functions/src/routes.ts
@@ -1,0 +1,6 @@
+import Router from 'express-promise-router'
+import * as slackCommandMessages from './apis/slack/command/messages'
+
+export const router = Router()
+
+router.use('/slack/command/messages', slackCommandMessages.router)
